@@ -14,29 +14,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.12",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.12",
-    "@angular/compiler": "^7.2.12",
-    "@angular/core": "^7.2.12",
-    "@angular/forms": "^7.2.12",
-    "@angular/http": "^7.2.12",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.12",
-    "@angular/platform-browser-dynamic": "7.2.12",
-    "@angular/router": "^7.2.12",
-    "core-js": "^3.0.1",
+    "@angular/animations": "^8.0.3",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "^8.0.3",
+    "@angular/compiler": "^8.0.3",
+    "@angular/core": "^8.0.3",
+    "@angular/forms": "^8.0.3",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.0.1",
+    "@angular/platform-browser": "^8.0.3",
+    "@angular/platform-browser-dynamic": "8.0.3",
+    "@angular/router": "^8.0.3",
+    "core-js": "^3.1.4",
     "hammerjs": "^2.0.8",
-    "rxjs": "^6.4.0",
-    "zone.js": "^0.9.0"
+    "rxjs": "^6.5.2",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular/cli": "^7.3.8",
-    "@angular/compiler-cli": "^7.2.12",
+    "@angular-devkit/build-angular": "^0.800.6",
+    "@angular/cli": "^8.0.6",
+    "@angular/compiler-cli": "^8.0.3",
     "@types/hammerjs": "^2.0.36",
     "@types/jasmine": "3.3.12",
-    "@types/node": "~11.13.2",
+    "@types/node": "~12.0.10",
     "codelyzer": "~5.0.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -49,8 +49,8 @@
     "karma-phantomjs-launcher": "^1.0.4",
     "phantomjs-prebuilt": "^2.1.16",
     "protractor": "^5.4.2",
-    "ts-node": "~8.0.3",
+    "ts-node": "~8.3.0",
     "tslint": "~5.15.0",
-    "typescript": "~3.2.4"
+    "typescript": "~3.4.3"
   }
 }
